Create @vanilla/utils
