+++ packages/vanilla-utils/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig-base.json",
+    "compilerOptions": {
+        "baseUrl": ".",
+        "lib": ["es2018", "dom"],
+        "paths": {
+            "@vanilla/utils/*": ["./src/*"]
+        }
+    }
+}
